pip-requirements.txt
Django==2.2.4
djangorestframework==3.10.2
freeze==1.0.10
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
